Introduce general Google_Site_Kit_Client base class for useful customizations.
